requirements.txt
django-taggit==1.2.0
Django==3.1
django-allauth==0.40.0
django-crispy-forms==1.9.1